msm: rq_stats: move away from tick-sched
IMHO tick-sched is the wrong place for this piece of code. The only thing
which needs to be done in there is to call (void update_rq_stats()).

Let's move update_rq_stats() and wakeup_user() into arch/arm/mach-msm/msm_rq_stats.c

Since wakeup_user() is called unconditionally and doesn't seem to be needed
elsewhere, get rid of this function and migrate it to update_rq_stats().

To avoid possible conflicts and to make clear where 'update_rq_stats' belongs
to we rename it into 'msm_update_rq_stats'.

On this opportunity we also define msm-rq-stats internal declarations as static.

And for the case we disable those rq-stats one day (we will for sure!) add
some ifdef's within tick-sched.c

Signed-off-by: moddingg33k <moddingg33k@t-online.de>