msm: smem: check if shared memory has been initialized
The shared memory protocol requires that the system bootloader initialize
the shared memory region before Linux boots to prevent race conditions.  As
the shared memory region is a vital part of the system, it is assumed that
dependency is always met.  There exists some corner cases in early target
development where that assumption does not hold.

Add a mechanism to check if the shared memory init assumption was violated
so that appropriate failsafe actions can be taken.

CRs-Fixed: 491395
Change-Id: Ife9bc71a405881f5c2945a8cd937db0f7db7e944
Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>

Signed-off-by: moddingg33k <moddingg33k@t-online.de>